cluster: initial commit
raft+boltdb to store clustered state

The overall scheme is as follows:
1. Typically, network is a bigger bottleneck than CPU and memory
2. Clustering to provide for a fatter network
3. The shared state requiring consensus is the "repo->node routing
table"
4. A client may reach any node in the cluster for a repo, if so lookup
routing table
        4.1 If data is local, serve the data
        4.2 If not, return a 301 redirect to the proper node
5. For additional fault-tolerance, repos may be placed on
highly-available storage clusters, or, we also do N-way replication.

TBD: should we allow starting life as single-node cluster and then grow
which would be best if nodes are on a shared storage cluster.

load-balancing will depend on the routing table

Best to implement this as a middleware at least for lookups.
Whenever the quorum changes, adjust routing table and commit.
We can assume storage is on a shared cluster at least initially.

example routing table
/a -> node1
/b -> node2
/c -> node3
/d -> node1
...

or equivalently,
node1 is responsible for [/a, /d ...], etc

Signed-off-by: Ramkumar Chinchani <rchincha@cisco.com>