feat/scale-out: support for proxying GQL queries
Previously, only dist-spec APIs were supported for scale-out as in a
shared storage environment, the metadata was shared and any instance
could correctly respond to the GQL queries as all the data is available.

In a local scale-out cluster deployment, the metadata store, in addition
to the file storage is isolated to each member in the cluster.
Due to this, there is a need to proxy the GQL queries as well for UI
and client requests to work as expected.

This change introduces a new GQL proxy + a handler for the GlobalSearch
request that proxies the request to all the members and collects them
for response to the client.
Support for other GQL queries is pending.

Signed-off-by: Vishwas Rajashekar <vrajashe@cisco.com>