chore: use a new build tag internal_test
This is to avoid drawing in unintended dependencies in the production binaries, since *internal_test.go files are in the same package with the production code.
This was the reason why we saw a spike in binary size during: https://github.com/project-zot/zot/actions/runs/19495162907/job/55795492766
This commit should fix it.

Signed-off-by: Andrei Aaron <andreifdaaron@gmail.com>